refactor: move router to main routing class
