+++ app/src/app/shared/components/search/search.component.html
@@ -29,7 +29,7 @@
       class="flex-initial flex align-items-center justify-content-end font-bold"
     >
       <div>
-        <a routerLink="/courses/add"
+        <a routerLink="/courses/new"
           ><p-button label="Добавить курс" icon="pi pi-plus"></p-button
         ></a>
       </div>
